+++ node/clients/index.ts
@@ -1,10 +1,9 @@
 import { IOClients } from '@vtex/api'
+import { masterDataFor } from '@vtex/clients'
+import type { Affiliates } from 'vtex.affiliates'
 
-import Status from './status'
-
-// Extend the default IOClients implementation with our own custom clients.
 export class Clients extends IOClients {
-  public get status() {
-    return this.getOrSet('status', Status)
+  public get affiliates() {
+    return this.getOrSet('affiliates', masterDataFor<Affiliates>('affiliates'))
   }
 }