+++ node/index.ts
@@ -1,17 +1,15 @@
 import type { ClientsConfig, ServiceContext, RecorderState } from '@vtex/api'
-import { LRUCache, method, Service } from '@vtex/api'
+import { method, Service } from '@vtex/api'
 
 import { Clients } from './clients'
-import { status } from './middlewares/status'
-import { validate } from './middlewares/validate'
+import { createAffiliate } from './middlewares/createAffiliate'
+import { updateAffiliate } from './middlewares/updateAffiliate'
+import { validateCreate } from './middlewares/validateCreate'
+import { validateUpdate } from './middlewares/validateUpdate'
+import { isAffiliateValid } from './resolvers/isAffiliateValid'
+import type { AffiliateInput } from './typings/affiliates'
 
-const TIMEOUT_MS = 800
-
-// Create a LRU memory cache for the Status client.
-// The @vtex/api HttpClient respects Cache-Control headers and uses the provided cache.
-const memoryCache = new LRUCache<string, any>({ max: 5000 })
-
-metrics.trackCache('status', memoryCache)
+const TIMEOUT_MS = 1000
 
 // This is the configuration for clients available in `ctx.clients`.
 const clients: ClientsConfig<Clients> = {
@@ -20,12 +18,8 @@ const clients: ClientsConfig<Clients> = {
   options: {
     // All IO Clients will be initialized with these options, unless otherwise specified.
     default: {
-      retries: 2,
+      retries: 3,
       timeout: TIMEOUT_MS,
-    },
-    // This key will be merged with the default options and add this cache to our Status client.
-    status: {
-      memoryCache,
     },
   },
 }
@@ -36,7 +30,7 @@ declare global {
 
   // The shape of our State object found in `ctx.state`. This is used as state bag to communicate between middlewares.
   interface State extends RecorderState {
-    code: number
+    affiliate?: AffiliateInput
   }
 }
 
@@ -44,9 +38,16 @@ declare global {
 export default new Service({
   clients,
   routes: {
-    // `status` is the route ID from service.json. It maps to an array of middlewares (or a single handler).
-    status: method({
-      GET: [validate, status],
+    affiliate: method({
+      POST: [validateCreate, createAffiliate],
+      PATCH: [validateUpdate, updateAffiliate],
     }),
+  },
+  graphql: {
+    resolvers: {
+      Query: {
+        isAffiliateValid,
+      },
+    },
   },
 })