+++ node/resolvers/isAffiliateValid.ts
@@ -1,0 +1,19 @@
+import { isSlugValid } from '../utils/shared'
+
+type Props = {
+  affiliateId: string
+}
+
+export const isAffiliateValid = async (
+  _: unknown,
+  { affiliateId }: Props,
+  { clients: { affiliates } }: Context
+) => {
+  if (!isSlugValid(affiliateId)) {
+    throw new Error('Affiliate ID is not valid')
+  }
+
+  const affiliateData = await affiliates.get(affiliateId, ['id', 'isApproved'])
+
+  return affiliateData?.isApproved ?? false
+}